Python_script/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
numpy==1.26.3
SimpleITK==2.3.1
Pillow==10.2.0
cachetools==5.3.2
python-multipart==0.0.6